questao03/pyt-decrypt/pyproject.toml
[tool.poetry]
name = "pyt-decrypt"
version = "0.1.0"
description = ""
authors = ["Filipe Silva <filipe.ufsc@gmail.com>"]
readme = "README.md"
packages = [{include = "pyt_decrypt"}]


[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.21.0"
typer = "^0.12.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
blue = "^0.9.1"
taskipy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.taskipy.tasks]
lint = "blue --check --diff ."
pre_test = "task lint"
test = "pytest -s -x --cov=pyt_decrypt -vv"
post_test = "coverage html"